Revert "Merge branch 'main' into Physics"
This reverts commit 1493e2db087162210ae8f04d764a9cfa6eac7ecb, reversing
changes made to 0c081d85fdc8e03099d9cf1b717d73416295c0e6.

+++ CarmicahEngine/Carmicah/source/InputSystem.cpp
@@ -73,7 +73,7 @@ namespace Carmicah
 					}
 				}
 			}
-			mMouseTick += static_cast<float>(CarmicahTimer::GetDt());
+			mMouseTick += CarmicahTimer::GetDt();
 		}	
 	}
